Add unit test for display_label using computed attribute
